requirements.txt
streamlit
gtts
googletrans==4.0.0-rc1
openai-whisper
pydub
